+++ src/Plugin/src/Export/IgcExport/src/IgcExportPlugin.cpp
@@ -59,49 +59,49 @@
 namespace
 {
     // General
-    constexpr char LineEnd[] = "\r\n";
-    constexpr char DateFormat[] = "ddMMyy";
-    constexpr char TimeFormat[] = "hhmmss";
+    constexpr char LineEnd[] {"\r\n"};
+    constexpr char DateFormat[] {"ddMMyy"};
+    constexpr char TimeFormat[] {"hhmmss"};
 
     // A record
-    constexpr char ManufacturerCode[] = "XXY";
-    constexpr char SerialId[] = "001";
+    constexpr char ManufacturerCode[] {"XXY"};
+    constexpr char SerialId[] {"001"};
 
     // H record
-    constexpr char Date[] = "FDTEDATE:";
-    constexpr char Pilot[] = "FPLTPILOTINCHARGE:";
-    constexpr char CoPilot[] = "FCM2CREW2:";
-    constexpr char GliderType[] = "FGTYGLIDERTYPE:";
-    constexpr char GliderId[] = "FGIDGLIDERID:";
-    constexpr char GPSDatum[] = "FDTMGPSDATUM:WGS84";
-    constexpr char FirmwareVersion[] = "FRFWFIRMWAREVERSION:";
-    constexpr char HardwareVersion[] = "FRHWHARDWAREVERSION:";
-    constexpr char FlightRecorderType[] = "FFTYFRTYPE:";
-    constexpr char GpsReceiver[] = "FGPSRECEIVER:";
-    constexpr char PressureAltitudeSensor[] = "FPRSPRESSALTSENSOR:";
-    constexpr char Security[] = "FFRSSECURITYOK";
+    constexpr char Date[] {"FDTEDATE:"};
+    constexpr char Pilot[] {"FPLTPILOTINCHARGE:"};
+    constexpr char CoPilot[] {"FCM2CREW2:"};
+    constexpr char GliderType[] {"FGTYGLIDERTYPE:"};
+    constexpr char GliderId[] {"FGIDGLIDERID:"};
+    constexpr char GPSDatum[] {"FDTMGPSDATUM:WGS84"};
+    constexpr char FirmwareVersion[] {"FRFWFIRMWAREVERSION:"};
+    constexpr char HardwareVersion[] {"FRHWHARDWAREVERSION:"};
+    constexpr char FlightRecorderType[] {"FFTYFRTYPE:"};
+    constexpr char GpsReceiver[] {"FGPSRECEIVER:"};
+    constexpr char PressureAltitudeSensor[] {"FPRSPRESSALTSENSOR:"};
+    constexpr char Security[] {"FFRSSECURITYOK"};
 
     // I record
-    constexpr char EnvironmentalNoiseLevel[] = "ENL";
+    constexpr char EnvironmentalNoiseLevel[] {"ENL"};
 
     // J record
-    constexpr char TrueHeading[] = "HDT";
-    constexpr char IndicatedAirspeed[] = "IAS";
+    constexpr char TrueHeading[] {"HDT"};
+    constexpr char IndicatedAirspeed[] {"IAS"};
 
     // C record
-    constexpr char ObsoleteFlightDate[] = "000000";
-    constexpr char ObsoleteTaskNumber[] = "000000";
-    constexpr char TakeoffPoint[] = "TAKEOFF";
-    constexpr char StartPoint[] = "START";
-    constexpr char TurnPoint[] = "TURN";
-    constexpr char FinishPoint[] = "FINISH";
-    constexpr char LandingPoint[] = "LANDING";
+    constexpr char ObsoleteFlightDate[] {"000000"};
+    constexpr char ObsoleteTaskNumber[] {"000000"};
+    constexpr char TakeoffPoint[] {"TAKEOFF"};
+    constexpr char StartPoint[] {"START"};
+    constexpr char TurnPoint[] {"TURN"};
+    constexpr char FinishPoint[] {"FINISH"};
+    constexpr char LandingPoint[] {"LANDING"};
 
     // B record
-    constexpr char FixValid[] = "A";
+    constexpr char FixValid[] {"A"};
 
     // Interval of 20 seconds for K records
-    constexpr int KRecordIntervalSec = 20;
+    constexpr int KRecordIntervalSec {20};
 }
 
 class IgcExportPluginPrivate
@@ -289,16 +289,21 @@ inline bool IgcExportPlugin::exportFixes(const Aircraft &aircraft, QIODevice &io
     QDateTime startTime = d->flight->getAircraftStartZuluTime(aircraft);
     QDateTime lastKFixTime;
 
+    Convert convert;
     const Engine &engine = aircraft.getEngineConst();
     std::vector<PositionData> interpolatedPositionData;
     Export::resamplePositionDataForExport(aircraft, d->pluginSettings.getResamplingPeriod(), std::back_inserter(interpolatedPositionData));
     bool ok = true;
     for (PositionData &positionData : interpolatedPositionData) {
         if (!positionData.isNull()) {
-            const int altitude = qRound(Convert::feetToMeters(positionData.altitude));
-            const QByteArray altitudeByteArray = formatNumber(altitude, 5);
-            const int indicatedAltitude = qRound(Convert::feetToMeters(positionData.indicatedAltitude));
-            const QByteArray indicatedAltitudeByteArray = formatNumber(indicatedAltitude, 5);
+
+            // Convert height above EGM geoid to height above WGS84 ellipsoid (HAE) [meters]
+            const double heightAboveEllipsoid = convert.egmToWgs84Ellipsoid(positionData.latitude, positionData.longitude, Convert::feetToMeters(positionData.altitude));
+
+            const int gnssAltitude = qRound(heightAboveEllipsoid);
+            const QByteArray gnssAltitudeByteArray = formatNumber(gnssAltitude, 5);
+            const int pressureAltitude = qRound(Convert::feetToMeters(positionData.indicatedAltitude));
+            const QByteArray pressureAltitudeByteArray = formatNumber(pressureAltitude, 5);
             const EngineData &engineData = engine.interpolate(positionData.timestamp, TimeVariableData::Access::Linear);
             const int noise = estimateEnvironmentalNoise(engineData);
             const QDateTime currentTime = startTime.addMSecs(positionData.timestamp);
@@ -307,9 +312,9 @@ inline bool IgcExportPlugin::exportFixes(const Aircraft &aircraft, QIODevice &io
                                        formatPosition(positionData.latitude, positionData.longitude) %
                                        ::FixValid %
                                        // Pressure altitude
-                                       indicatedAltitudeByteArray %
+                                       pressureAltitudeByteArray %
                                        // GNSS altitude
-                                       altitudeByteArray %
+                                       gnssAltitudeByteArray %
                                        formatNumber(noise, 3) %
                                        ::LineEnd;
             ok = io.write(bRecord);
@@ -367,7 +372,7 @@ inline QByteArray IgcExportPlugin::IgcExportPlugin::formatLatitude(double latitu
 {
     int degrees;
     double minutes;
-    Convert::dd2dm(latitude, degrees, minutes);
+    Convert::ddTodm(latitude, degrees, minutes);
     const int decimals = static_cast<int>((minutes - static_cast<int>(minutes)) * 1000);
     const QByteArray latitudeString = QString("%1%2%3%4")
             .arg(degrees, 2, 10, QLatin1Char('0'))
@@ -382,7 +387,7 @@ inline QByteArray IgcExportPlugin::IgcExportPlugin::formatLongitude(double longi
 {
     int degrees;
     double minutes;
-    Convert::dd2dm(longitude, degrees, minutes);
+    Convert::ddTodm(longitude, degrees, minutes);
     const int decimals = static_cast<int>((minutes - static_cast<int>(minutes)) * 1000);
     const QByteArray latitudeString = QString("%1%2%3%4")
             .arg(degrees, 3, 10, QLatin1Char('0'))